fix: removed pooler/sampler distinction
