--- src/autora/experimentalist/pooler/example_pooler/__init__.py
@@ -1,20 +1,0 @@
-"""
-Example Experimentalist Sampler
-"""
-def example_pool(argument: float) -> float:
-    """
-    Add a description of the pooler here
-
-    Args:
-        argument: description of the argument
-    Returns: pool of experimental conditions
-
-    *Optional*
-    Examples:
-        These examples add documentations and also work as doctests
-        >>> example_pool(1.)
-        1.0
-    """
-    new_conditions = argument
-
-    return new_conditions
